remove support for build_environment_vars_for's deactivate option
When the deactivate option was added, it changed the signature of the
build_environment_vars_for method incompatibly.  This broke the CPAN.pm
bootstrapping process.  The deactivate option doesn't appear to be used
anywhere on CPAN except in error.

Since we don't need the interpolate option anymore, we can just ignore
any extra parameters.  This will fix CPAN.pm.